Align product name with button
This uses the technique described in this SO post  http://stackoverflow.com/questions/20547819/vertical-align-with-bootstrap-3.

Special cases the margin on the bottom of the paragraph tag - we may not
need to use a p tag at all, but I'm so unfamiliar with how all this works
I've popped it in there.