Changes sign up form text
Now says "... to create an account" instead of "... to sign in!" (made it less shouty too).

Also, remove the show-grid class from the Row as it didn't seem to have any effect.

Finally, removed an unused import and added the necessary to package.json to enable npm run deploy HOWEVER!!! you still need to put in a homepage setting (but this will be different for everyone and using environment variables didn't seem to be that straight forward).

So, to deploy to GitHub pages do the following:

- add '"homepage": "http://<username>.github.io/web",
- run 'npm run build' to do a release build
- run 'npm run deploy' to deploy to GitHub pages (you shouldn't need to do the install for the gh-pages node module but you will need to run npm install once just to make it available to you... I think).