Adds Buy button to BuyScreen
In addition this uses its own state to decide if the user is signed up or not. If not, the SignUpForm is shown but the TopUpForm is not. The Buy button is grey and deactivated. If the user is signed up it shows a zero balance, displays the TopUpForm and activates and makes green the Buy button.

No way currently to change the state.