Hooks up email Submit button
Still not calling any API but just storing stuff locally for now.

On submitting an email address a balance of £0.00 is shown and the TopUpForm is displayed. The Buy button is still greyed out because the TopUpForm won't yet increase the balance (but would be green if balance was sufficient).

The Buy button itself has been extracted into its own Component.