Handles rejected top ups
This I think closes #6.

I've turned the ProductStore into a MockApi so we can pretend to make requests on it. To be fair it's a mixture of API and state so would need to be restructured but it allows us to get going for now, i.e. I realise it's pretty naff!

The BuyScreen now checks the card details via the MockApi and if they're successful will go to the Succesful purchase screen when that's done. If failed, it pops up a Modal error message which I've done as a separate component.