Add api routing for registration stage
Using `/register` the user has an access and refresh token generated
using the 'jsonwebtoken' package. This is then sent as part of the
response.

`register2` is passed the access token as part of the authorization
header. This token is verified before any further evaluation of the
card details/item ID. If all is well, a new transaction is added to
the userID -> transactions map and the JSON response returns the user's
balance and transaction history.