Add 'http-status' package and update status codes
When adding this package and doing a bit of googling, it seems like
the 403 status code (forbidden) was the incorrect one to use.

'Forbidden' should be used when the user can access the endpoint but
does not have permission. In our case, we want to indicate the user has
not been authenticated. See http://stackoverflow.com/a/3297070 for more
info.