don't create a (transaction) account in register
Accounts are only created by the user service when we successfully
register(2) a user. Previously we'd create an account in register,
for things like getSessionData(), but this account would be overwritten
by the one from register2. This meant the latter took precedence
and would usually end up with all the transactions, however in cases
of a CDN caching the initial user JSON with the initial accountId,
our custom topup script would credit the initial accountId with
fund transfers, from e.g. scottcoin.