api handles a registered user without card details
Previously api would allow an exception to propagate to the frontend
UI, meaning if a user registered but their card topup didn't work,
they'd be permanently stuck on an error screen.

Now if they successfully register (but fail to topup) they can still
browse the store and more importantly, submit support questions.

The topup page will still fail / raise an error locally in the UI,
but this is fixed by another commit.