Show user store after successful registration
This was relying on a hack which no longer works - the fact that we
registered a user twice.

On the second attempt to navigate to '/store' a refreshToken existed
which enabled us to show the loading page.

We now only show the store when the 1st phase of registration was
successful. This breaks the loading animation, however this needs to be
pulled out to a higher level than page anyway so we can show the loading
animation for the Home component as well.