always attempt to use the auth'd request-key before extracting/creating a service key
Previously we would ignore `request.key` on a service route, instead
attempting to either extract an existing key (which worked fine for
sub-services) or create one (which is what the api service was
doing).

The problem is that api service would have authenticated users and
the `request` object would already have a key on it, which we were
ignoring. We now always attempt to use this key (on which there is
additional information) before attempting to extract a key from the
request headers or create a service key.