move box to success page before dispatching success
Not sure if this is the correct order all our dispatch and history
pushes should be. Main cause of this problem is that we fire off a
request as soon as the component is rendered. Dispatching success
results in the component being re-created which will then cause an
extra call to be made to the API.

If we push the page onto the success page before dispatching success
then we don't re-render our BoxReceived component.