use cruft's reducer to prevent duplicate slack notifications
Some things to note:
	1) aggregate is either 'refund' or 'purchase'
	2) we're not using cruft's emit() as we can only emit other database entries, not slack messages

Unfortunately #2 means that we have to use an (arguably hacky) way
of figuring out if we've already reduced the current event or not.

Additionally, if the slack send subsequently fails, we'll have
already reduced said event and so won't emit another slack message
because future reduce()s of it will be no-ops.

This is perhaps fine for slack messages, but it's certainly not
something we can rely on (where "it" == "eventStored").

In short: this can be viewed as a quick fix for the #purchases
channel, but I fear leaving it in long time will cause us to forget
and misuse the return of reduce()().